+++ .gitignore
@@ -1,0 +1,3 @@
+**/__*__
+**/venv
+**/.env
