+++ db/models.py
@@ -1,0 +1,10 @@
+from sqlalchemy import Boolean, Column, Integer, String
+
+from db import database
+
+
+class Word(database.Base):
+    __tablename__ = "words"
+
+    frequency = Column(Integer, primary_key=False)
+    word = Column(String, unique=False, primary_key=True)
